requirements.txt
autograd==1.5
black==23.3.0
click==8.1.3
cma==2.7.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
fonttools==4.39.4
future==0.18.3
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.40.0
matplotlib==3.7.1
mypy-extensions==1.0.0
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pygco==0.0.16
pymoo==0.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
